move hook logic to context provider; introduce useCallbacks
