+++ packages/abstraxion/src/hooks/useAbstraxionAccount.ts
@@ -1,6 +1,5 @@
-import { useContext, useEffect } from "react";
+import { useContext } from "react";
 import { AbstraxionContext } from "@/src/components/AbstraxionContext";
-import { abstraxionAuth } from "../components/Abstraxion";
 
 export interface AbstraxionAccount {
   bech32Address: string;
@@ -13,43 +12,8 @@ export interface AbstraxionAccountState {
 }
 
 export const useAbstraxionAccount = (): AbstraxionAccountState => {
-  const {
-    isConnected,
-    granterAddress,
-    abstraxionAccount,
-    isConnecting,
-    setIsConnected,
-    setAbstraxionAccount,
-    setGranterAddress,
-  } = useContext(AbstraxionContext);
-
-  useEffect(() => {
-    const unsubscribe = abstraxionAuth.subscribeToAuthStateChange(
-      async (newState: boolean) => {
-        if (Boolean(newState) === true) {
-          const account = await abstraxionAuth.getLocalKeypair();
-          const granterAddress = abstraxionAuth.getGranter();
-          setAbstraxionAccount(account);
-          setGranterAddress(granterAddress);
-        }
-        setIsConnected(newState);
-      },
-    );
-
-    return () => {
-      unsubscribe?.();
-    };
-  }, [abstraxionAuth]);
-
-  useEffect(() => {
-    async function persistAuthenticateState() {
-      await abstraxionAuth.authenticate();
-    }
-
-    if (!isConnecting && !abstraxionAccount && !granterAddress) {
-      persistAuthenticateState();
-    }
-  }, [isConnected, abstraxionAccount, granterAddress, abstraxionAuth]);
+  const { isConnected, granterAddress, isConnecting } =
+    useContext(AbstraxionContext);
 
   return {
     data: {
